requirements.txt
garminconnect>=0.2.19
python-dotenv>=1.0.0
pyyaml>=6.0
